# web/openapi.json
{
  "components": {
    "schemas": {
      "Change": {
        "additionalProperties": false,
        "properties": {
          "item_id": {
            "maximum": 2147483647.0,
            "minimum": 1.0,
            "title": "Item Id",
            "type": "integer"
          },
          "value": {
            "enum": [
              -1,
              0,
              1
            ],
            "title": "Value",
            "type": "integer"
          }
        },
        "required": [
          "item_id",
          "value"
        ],
        "title": "Change",
        "type": "object"
      },
      "Consent": {
        "additionalProperties": false,
        "properties": {
          "adult": {
            "const": true,
            "title": "Adult",
            "type": "boolean"
          },
          "research": {
            "const": true,
            "title": "Research",
            "type": "boolean"
          }
        },
        "required": [
          "adult",
          "research"
        ],
        "title": "Consent",
        "type": "object"
      },
      "Event": {
        "additionalProperties": false,
        "properties": {
          "event_id": {
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Event Id",
            "type": "string"
          },
          "event_time_ms": {
            "minimum": 0.0,
            "title": "Event Time Ms",
            "type": "integer"
          },
          "item_id": {
            "maximum": 2147483647.0,
            "minimum": 1.0,
            "title": "Item Id",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "shown",
              "save",
              "dismiss",
              "watched"
            ],
            "title": "Kind",
            "type": "string"
          },
          "request_id": {
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "request_id",
          "item_id",
          "kind",
          "event_time_ms"
        ],
        "title": "Event",
        "type": "object"
      },
      "EventResult": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          }
        },
        "required": [
          "accepted",
          "duplicate"
        ],
        "title": "EventResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Me": {
        "additionalProperties": false,
        "properties": {
          "consent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Version"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "preference_revision": {
            "title": "Preference Revision",
            "type": "integer"
          },
          "required_consent": {
            "title": "Required Consent",
            "type": "string"
          }
        },
        "required": [
          "id",
          "consent_version",
          "required_consent",
          "preference_revision"
        ],
        "title": "Me",
        "type": "object"
      },
      "Movie": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "title": "Genres",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "title",
          "year",
          "genres",
          "available"
        ],
        "title": "Movie",
        "type": "object"
      },
      "MoviePage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Movie"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "MoviePage",
        "type": "object"
      },
      "PreferencePage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Change"
            },
            "title": "Items",
            "type": "array"
          },
          "preference_revision": {
            "title": "Preference Revision",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "preference_revision"
        ],
        "title": "PreferencePage",
        "type": "object"
      },
      "PreferenceResult": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "preference_revision": {
            "title": "Preference Revision",
            "type": "integer"
          }
        },
        "required": [
          "preference_revision",
          "count"
        ],
        "title": "PreferenceResult",
        "type": "object"
      },
      "Preferences": {
        "additionalProperties": false,
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/Change"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Changes",
            "type": "array"
          },
          "request_id": {
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "changes"
        ],
        "title": "Preferences",
        "type": "object"
      },
      "Recommendation": {
        "additionalProperties": false,
        "properties": {
          "k": {
            "default": 10,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "K",
            "type": "integer"
          },
          "request_id": {
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "Recommendation",
        "type": "object"
      },
      "RecommendationResult": {
        "additionalProperties": false,
        "properties": {
          "candidate_source": {
            "enum": [
              "popularity",
              "centroid",
              "als"
            ],
            "title": "Candidate Source",
            "type": "string"
          },
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "exhausted": {
            "title": "Exhausted",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecommendedMovie"
            },
            "title": "Items",
            "type": "array"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          },
          "preference_revision": {
            "title": "Preference Revision",
            "type": "integer"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "model_version",
          "policy_version",
          "preference_revision",
          "candidate_source",
          "degraded",
          "exhausted",
          "items"
        ],
        "title": "RecommendationResult",
        "type": "object"
      },
      "RecommendedMovie": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "title": "Genres",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "source": {
            "enum": [
              "popularity",
              "centroid",
              "als"
            ],
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "title",
          "year",
          "genres",
          "available",
          "source"
        ],
        "title": "RecommendedMovie",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WatchedMovie": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "dismissed": {
            "title": "Dismissed",
            "type": "boolean"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "title": "Genres",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "saved": {
            "title": "Saved",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "watched": {
            "title": "Watched",
            "type": "boolean"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "title",
          "year",
          "genres",
          "available",
          "saved",
          "watched",
          "dismissed"
        ],
        "title": "WatchedMovie",
        "type": "object"
      },
      "Watchlist": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WatchedMovie"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "Watchlist",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "RecServe movie pilot",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v2/events": {
      "post": {
        "operationId": "feedback_api_v2_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feedback"
      }
    },
    "/api/v2/me": {
      "delete": {
        "operationId": "remove_api_v2_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Remove"
      },
      "get": {
        "operationId": "me_api_v2_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Me"
      }
    },
    "/api/v2/me/consent": {
      "put": {
        "operationId": "consent_api_v2_me_consent_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Consent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consent"
      }
    },
    "/api/v2/me/export": {
      "get": {
        "operationId": "export_api_v2_me_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Export"
      }
    },
    "/api/v2/movies": {
      "get": {
        "operationId": "movies_api_v2_movies_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 100,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoviePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Movies"
      }
    },
    "/api/v2/preferences": {
      "get": {
        "operationId": "get_preferences_api_v2_preferences_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencePage"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Preferences"
      },
      "put": {
        "operationId": "set_preferences_api_v2_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Preferences"
      }
    },
    "/api/v2/recommendations": {
      "post": {
        "operationId": "recommendations_api_v2_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Recommendation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recommendations"
      }
    },
    "/api/v2/watchlist": {
      "get": {
        "operationId": "watchlist_api_v2_watchlist_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Watchlist"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Watchlist"
      }
    },
    "/auth/callback": {
      "get": {
        "operationId": "callback_auth_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "",
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "default": "",
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Callback"
      }
    },
    "/auth/login": {
      "get": {
        "operationId": "login_auth_login_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Login"
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Logout"
      }
    },
    "/healthz": {
      "get": {
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/readyz": {
      "get": {
        "operationId": "ready_readyz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready"
      }
    }
  }
}